adding test 9 for formatDate function
